Fix save/load value

+++ src/texttospeech/texttospeechconfigwidget.cpp
@@ -105,8 +105,8 @@ void TextToSpeechConfigWidget::readConfig()
 {
     KConfig config(QStringLiteral("texttospeechrc"));
     KConfigGroup grp = config.group("Settings");
-    mRate->setValue(static_cast<int>(grp.readEntry("rate", 0) * 100));
-    mPitch->setValue(static_cast<int>(grp.readEntry("pitch", 0) * 100));
+    mRate->setValue(static_cast<int>(grp.readEntry("rate", 0.0) * 100));
+    mPitch->setValue(static_cast<int>(grp.readEntry("pitch", 0.0) * 100));
     mVolume->setValue(static_cast<int>(grp.readEntry("volume", 50)));
     updateLocale();
 }
@@ -116,8 +116,8 @@ void TextToSpeechConfigWidget::writeConfig()
     KConfig config(QStringLiteral("texttospeechrc"));
     KConfigGroup grp = config.group("Settings");
     grp.writeEntry("volume", mVolume->value());
-    grp.writeEntry("rate", static_cast<double>(mRate->value() / 100));
-    grp.writeEntry("pitch", static_cast<double>(mPitch->value() / 100));
+    grp.writeEntry("rate", (static_cast<int>((static_cast<double>(mRate->value()) / 100.0))));
+    grp.writeEntry("pitch", (static_cast<int>((static_cast<double>(mPitch->value()) / 100.0))));
     grp.writeEntry("localeName", mLanguage->currentData().toLocale().name());
     grp.writeEntry("engine", mAvailableEngine->currentData().toString());
     grp.writeEntry("voice", mVoice->currentData().toString());
